move 'lexer_from_source()' and 'dict_to_pytest_param()' to tests/utils.py
